requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
rsa==4.9
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
